+++ mod/racing_plus/SpeedrunPostRender.lua
@@ -206,6 +206,11 @@ function SpeedrunPostRender:DrawVetoButtonText()
 end
 
 function SpeedrunPostRender:DrawSeason7Goals()
+  local challenge = Isaac.GetChallenge()
+  if challenge ~= Isaac.GetChallengeIdByName("R+7 (Season 7 Beta)") then
+    return
+  end
+
   -- Only show the remaining goals if the user is pressing tab
   local tabPressed = false
   for i = 0, 3 do -- There are 4 possible inputs/players from 0 to 3