+++ mod/CHANGES.md
@@ -116,6 +116,7 @@ While there are thousands of rooms in the game, many players have already seen t
 * The Forsaken is now immune to fear. (This prevents the bug where it will not attack.)
 * Blastocyst is now immmune to freeze. (This prevents delays during the death animation.)
 * The "Would you like to do a Victory Lap!?" popup no longer appears after defeating The Lamb.
+* All pills can now be used to cancel pedestal pickup animations.
 
 ### Cutscene & Animation Removal